requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-model-utils==3.0.0
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.6
jsonfield==2.0.2
pytz==2017.2
requests==2.18.4
urllib3==1.22
python-crontab==2.2.4
